(PC-6877) feat:users: add repos functions
Add functions to: 1) find favorites that can be used to send a
notification (offer not expired nor booked by the user, still available,
no notification sent) ; 2) find for each user the number of bookings for
each categories.

Secod function will be needed to filter favorites from the first one:
for each user if more than one is found, keep the one that comes from an
offer with the lowest number of bookings.

These two functions will be used to send notifications: fetch the
favorites (1), filter them to keep only one by user max (2), send
notifications to encourage the user to book the offer.